Move asm_lang out of the language module.
