+++ sway-core/src/asm_generation/allocated_abstract_instruction_set.rs
@@ -1,4 +1,4 @@
-use crate::language::asm_lang::{
+use crate::asm_lang::{
     allocated_ops::{AllocatedOpcode, AllocatedRegister},
     AllocatedAbstractOp, ConstantRegister, ControlFlowOp, Label, RealizedOp, VirtualImmediate12,
     VirtualImmediate18, VirtualImmediate24,
