+++ sway-core/src/asm_generation/asm_builder/functions.rs
@@ -5,11 +5,11 @@ use super::{
 
 use crate::{
     asm_generation::{from_ir::*, Entry},
-    error::*,
-    language::asm_lang::{
+    asm_lang::{
         virtual_register::*, Op, OrganizationalOp, VirtualImmediate12, VirtualImmediate24,
         VirtualOp,
     },
+    error::*,
 };
 
 use sway_ir::*;
