+++ sway-core/src/asm_lang/virtual_immediate.rs
