+++ sway-core/src/asm_lang/virtual_ops.rs
