+++ sway-core/src/asm_lang/virtual_register.rs
