+++ sway-core/src/lib.rs
@@ -2,6 +2,7 @@
 pub mod error;
 
 mod asm_generation;
+mod asm_lang;
 mod build_config;
 mod concurrent_slab;
 pub mod constants;
